add connection error if network is absent

+++ app/src/main/java/com/example/lsdchat/util/UsersUtil.java
@@ -2,16 +2,16 @@ package com.example.lsdchat.util;
 
 
 import android.content.Context;
-import android.util.Log;
 
 import com.example.lsdchat.App;
 import com.example.lsdchat.api.dialog.model.ItemUser;
 import com.example.lsdchat.api.login.model.LoginUser;
 import com.example.lsdchat.manager.DataManager;
 import com.example.lsdchat.model.ContentModel;
-import com.example.lsdchat.ui.main.fragment.BaseFragment;
 import com.example.lsdchat.ui.main.users.UsersContract;
 import com.example.lsdchat.ui.main.users.UsersModel;
+import com.example.lsdchat.util.error.ErrorInterface;
+import com.example.lsdchat.util.error.NetworkConnect;
 
 import java.util.List;
 
@@ -20,38 +20,40 @@ import rx.android.schedulers.AndroidSchedulers;
 
 public class UsersUtil {
 
-    static ErrorInterface errorInterface = new BaseFragment();
 
     public static void getUserListAndSave(String token, Context context) {
+        NetworkConnect networkConnect = ((NetworkConnect) context);
+        ErrorInterface errorInterface = ((ErrorInterface) context);
         UsersContract.Model mModel = new UsersModel(App.getSharedPreferencesManager(context),
-                App.getDataManager(),App.getApiManager().getDialogService());
+                App.getDataManager(), App.getApiManager().getDialogService());
+        if (networkConnect.isNetworkConnect()) {
+            mModel.getUserList(token)
+                    .subscribe(userListResponse -> {
+                        List<ItemUser> itemUsers = userListResponse.getItemUserList();
+                        if (userListResponse.getTotalEntries() < (mModel.getUsersQuick().size() + 2)) {
+                            mModel.deleteAllUSerQuick();
+                            saveUserToDb(itemUsers, mModel);
+                        } else {
+                            saveUserToDb(itemUsers, mModel);
+                        }
+                        getImage(itemUsers, token, context);
 
-        mModel.getUserList(token)
-                .subscribe(userListResponse -> {
-                    List<ItemUser> itemUsers = userListResponse.getItemUserList();
-                    if (userListResponse.getTotalEntries() < (mModel.getUsersQuick().size() + 2)) {
-                        mModel.deleteAllUSerQuick();
-                        saveUserToDb(itemUsers, mModel);
-                    } else {
-                        saveUserToDb(itemUsers, mModel);
-                    }
-                    getImage(itemUsers, token);
-
-                }, throwable -> Log.e("FIX", throwable.getMessage()));
+                    }, errorInterface::showErrorDialog);
+        }
     }
 
-    private static void getImage(List<ItemUser> itemUsers, String token) {
+    private static void getImage(List<ItemUser> itemUsers, String token, Context context) {
         DataManager dataManager = App.getDataManager();
-
+        ErrorInterface errorInterface = ((ErrorInterface) context);
         Observable.from(itemUsers)
                 .flatMap(user -> Observable.just(user.getUser()))
                 .filter(user -> user.getBlobId() != 0)
-                .doOnError(throwable -> errorInterface.showErrorDialog(throwable))
+                .doOnError(errorInterface::showErrorDialog)
                 .subscribe(user ->
                         Utils.downloadImage(user.getBlobId(), token)
-                        .flatMap(file -> Observable.just(file.getAbsolutePath()))
-                        .subscribe(path -> dataManager.saveUserAvatar(new ContentModel(String.valueOf(user.getId()),path)),
-                                throwable -> Log.e("FIX-user", throwable.getMessage())));
+                                .flatMap(file -> Observable.just(file.getAbsolutePath()))
+                                .subscribe(path -> dataManager.saveUserAvatar(new ContentModel(String.valueOf(user.getId()), path)),
+                                        errorInterface::showErrorDialog));
 
 
     }
